requirements.txt
apache-beam[gcp]==2.49.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
requests>=2.28.0
great_expectations>=0.16.0
apache-beam==2.49.0
